pyproject.toml
[project]
name = "mcp-incident-tool"
version = "0.1.0"
description = "Model Context Protocol Server for Automated Linux Crash Diagnosis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "asyncssh>=2.14.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0"
]

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]

[tool.setuptools]
packages = ["core", "parsers", "tools"]
py-modules = ["server", "conftest"]
